+++ TidalPy/utilities/__init__.py
@@ -1,2 +1,3 @@
 from .classes.config.dictionary_utils import nested_get
 from .io.progress import progress_bar
+from .performance.numba import use_numba