+++ TidalPy/utilities/performance/__init__.py
@@ -1,2 +1,2 @@
 from .cache import clear_cache
-from .numba import njit
+from .numba import njit, use_numba